% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_functions}
\alias{get_functions}
\title{Wrapper for some frequently used APIs to get data easily}
\description{
The main purpose is to handle APIs that have limitations on the number of results returned per single request.
}
\seealso{
\code{\link{get_positions_history}}
\code{\link{get_history_candles}}
}
