% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restAPI.R
\name{restAPIaccount}
\alias{restAPIaccount}
\title{restAPIaccount Class}
\description{
Wrapper for \href{https://www.okx.com/docs-v5/en/#rest-api-account}{REST API ACCOUNT}.
}
\section{Super class}{
\code{\link[okxAPI:restAPI]{okxAPI::restAPI}} -> \code{restAPIaccount}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-restAPIaccount-balance}{\code{restAPIaccount$balance()}}
\item \href{#method-restAPIaccount-positions}{\code{restAPIaccount$positions()}}
\item \href{#method-restAPIaccount-positions_history}{\code{restAPIaccount$positions_history()}}
\item \href{#method-restAPIaccount-clone}{\code{restAPIaccount$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="okxAPI" data-topic="restAPI" data-id="get_body"><a href='../../okxAPI/html/restAPI.html#method-restAPI-get_body'><code>okxAPI::restAPI$get_body()</code></a></span></li>
<li><span class="pkg-link" data-pkg="okxAPI" data-topic="restAPI" data-id="get_header"><a href='../../okxAPI/html/restAPI.html#method-restAPI-get_header'><code>okxAPI::restAPI$get_header()</code></a></span></li>
<li><span class="pkg-link" data-pkg="okxAPI" data-topic="restAPI" data-id="get_message"><a href='../../okxAPI/html/restAPI.html#method-restAPI-get_message'><code>okxAPI::restAPI$get_message()</code></a></span></li>
<li><span class="pkg-link" data-pkg="okxAPI" data-topic="restAPI" data-id="get_request_path"><a href='../../okxAPI/html/restAPI.html#method-restAPI-get_request_path'><code>okxAPI::restAPI$get_request_path()</code></a></span></li>
<li><span class="pkg-link" data-pkg="okxAPI" data-topic="restAPI" data-id="get_result"><a href='../../okxAPI/html/restAPI.html#method-restAPI-get_result'><code>okxAPI::restAPI$get_result()</code></a></span></li>
<li><span class="pkg-link" data-pkg="okxAPI" data-topic="restAPI" data-id="get_signature"><a href='../../okxAPI/html/restAPI.html#method-restAPI-get_signature'><code>okxAPI::restAPI$get_signature()</code></a></span></li>
<li><span class="pkg-link" data-pkg="okxAPI" data-topic="restAPI" data-id="get_timestamp"><a href='../../okxAPI/html/restAPI.html#method-restAPI-get_timestamp'><code>okxAPI::restAPI$get_timestamp()</code></a></span></li>
<li><span class="pkg-link" data-pkg="okxAPI" data-topic="restAPI" data-id="initialize"><a href='../../okxAPI/html/restAPI.html#method-restAPI-initialize'><code>okxAPI::restAPI$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-restAPIaccount-balance"></a>}}
\if{latex}{\out{\hypertarget{method-restAPIaccount-balance}{}}}
\subsection{Method \code{balance()}}{
See \href{https://www.okx.com/docs-v5/en/#rest-api-account-get-balance}{Get balance} for more information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{restAPIaccount$balance(ccy, process = "identity")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ccy}}{Single currency or a vector composed of multiple currencies. (no more than 20).}

\item{\code{process}}{A function to process the data received from the API. Default to \code{identity}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-restAPIaccount-positions"></a>}}
\if{latex}{\out{\hypertarget{method-restAPIaccount-positions}{}}}
\subsection{Method \code{positions()}}{
See \href{https://www.okx.com/docs-v5/en/#rest-api-account-get-positions}{Get positions} for more information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{restAPIaccount$positions(process = "identity", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{process}}{A function to process the data received from the API. Default to \code{identity}.}

\item{\code{...}}{Other request parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-restAPIaccount-positions_history"></a>}}
\if{latex}{\out{\hypertarget{method-restAPIaccount-positions_history}{}}}
\subsection{Method \code{positions_history()}}{
See \href{https://www.okx.com/docs-v5/en/#rest-api-account-get-positions-history}{Get positions history} for more information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{restAPIaccount$positions_history(process = "identity", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{process}}{A function to process the data received from the API. Default to \code{identity}.}

\item{\code{...}}{Other request parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-restAPIaccount-clone"></a>}}
\if{latex}{\out{\hypertarget{method-restAPIaccount-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{restAPIaccount$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
